Allow number query parameters to accept int or float

+++ src/Component/OpenApi2/Tests/fixtures/parameters/expected/Client.php
@@ -40,7 +40,7 @@ class Client extends \Jane\Component\OpenApi2\Tests\Expected\Runtime\Client\Clie
      * @param array $queryParameters {
      *     @var string $testString
      *     @var int $testInteger
-     *     @var float $testFloat
+     *     @var int $testFloat
      *     @var array $testArray
      *     @var string $testRequired
      *     @var string $testDefault
@@ -57,7 +57,7 @@ class Client extends \Jane\Component\OpenApi2\Tests\Expected\Runtime\Client\Clie
      * @param array $headerParameters {
      *     @var string $testString
      *     @var int $testInteger
-     *     @var float $testFloat
+     *     @var int $testFloat
      *     @var array $testArray
      *     @var string $testRequired
      *     @var string $testDefault
@@ -74,7 +74,7 @@ class Client extends \Jane\Component\OpenApi2\Tests\Expected\Runtime\Client\Clie
      * @param array $formParameters {
      *     @var string $testString
      *     @var int $testInteger
-     *     @var float $testFloat
+     *     @var int $testFloat
      *     @var array $testArray
      *     @var string $testRequired
      *     @var string $testDefault
@@ -112,12 +112,12 @@ class Client extends \Jane\Component\OpenApi2\Tests\Expected\Runtime\Client\Clie
     /**
      * @param string $testString
      * @param int $testInteger
-     * @param float $testFloat
+     * @param int $testFloat
      * @param string $fetch Fetch mode to use (can be OBJECT or RESPONSE)
      *
      * @return ($fetch is 'object' ? null : \Psr\Http\Message\ResponseInterface)
      */
-    public function testPathParameters(string $testString, int $testInteger, float $testFloat, string $fetch = self::FETCH_OBJECT)
+    public function testPathParameters(string $testString, int $testInteger, int $testFloat, string $fetch = self::FETCH_OBJECT)
     {
         return $this->executeEndpoint(new \Jane\Component\OpenApi2\Tests\Expected\Endpoint\TestPathParameters($testString, $testInteger, $testFloat), $fetch);
     }
